dirty libmodbus branch
So I decided to do a easy copy and paste work from libmodbus to fork a
new dynamixel c-api.

libmodbus more or less already includes all functions I needed so I
asked the author (Stéphane Raimbault) for permission to copy most of his
code. Thanks again!

This is the first commit - I already added the basics of the Dynamixel
TTL-protocol but it's totally untestet at the moment.